refactor: moved routing components to own high-level folder
