+++ src/components/AlertStatus/AlertStatus.tsx
@@ -4,10 +4,10 @@ import { Icon, IconButton, LinkButton, Tooltip, useStyles2, useTheme2 } from '@g
 import { css, cx } from '@emotion/css';
 
 import { AlertSensitivity, Check, PrometheusAlertsGroup, ROUTES } from 'types';
+import { getRoute } from 'routing/utils';
 import { useAlertRules } from 'hooks/useAlertRules';
 import { useMetricsDS } from 'hooks/useMetricsDS';
 import { AlertSensitivityBadge } from 'components/AlertSensitivityBadge';
-import { getRoute } from 'components/Routing.utils';
 import { Toggletip } from 'components/Toggletip';
 
 type AlertStatusProps = {
