+++ src/components/CheckEditor/CheckProbes/PrivateProbesAlert.tsx
@@ -3,7 +3,7 @@ import { Alert, LinkButton, Stack, TextLink } from '@grafana/ui';
 import { useLocalStorage } from 'usehooks-ts';
 
 import { ROUTES } from 'types';
-import { getRoute } from 'components/Routing.utils';
+import { getRoute } from 'routing/utils';
 
 export const PrivateProbesAlert = () => {
   const [dismissed, setDismissed] = useLocalStorage<boolean>('dismissedPrivateProbesAlert', false);
