+++ src/components/ConfigActions.tsx
@@ -4,10 +4,10 @@ import { Button, LinkButton } from '@grafana/ui';
 
 import { ROUTES } from 'types';
 import { hasGlobalPermission } from 'utils';
+import { getRoute } from 'routing/utils';
 import { useMeta } from 'hooks/useMeta';
 
 import { DisablePluginModal } from './DisablePluginModal';
-import { getRoute } from './Routing.utils';
 
 export const ConfigActions = ({ initialized }: { initialized?: boolean }) => {
   const [showDisableModal, setShowDisableModal] = useState(false);
