+++ src/components/OverLimitAlert/BrowserCheckLimitAlert.tsx
@@ -2,8 +2,8 @@ import React from 'react';
 import { TextLink } from '@grafana/ui';
 
 import { ROUTES } from 'types';
+import { getRoute } from 'routing/utils';
 import { useLimits } from 'hooks/useLimits';
-import { getRoute } from 'components/Routing.utils';
 import { Ul } from 'components/Ul';
 
 import { AlertContainer } from './AlertContainer';
