+++ src/components/OverLimitAlert/ExecutionLimitAlert.tsx
@@ -2,8 +2,8 @@ import React from 'react';
 import { TextLink } from '@grafana/ui';
 
 import { ROUTES } from 'types';
+import { getRoute } from 'routing/utils';
 import { FREE_EXECUTION_LIMIT } from 'hooks/useAtHgExecutionLimit';
-import { getRoute } from 'components/Routing.utils';
 import { Ul } from 'components/Ul';
 
 import { AlertContainer } from './AlertContainer';
