+++ src/components/ProbeCard/ProbeCard.tsx
@@ -2,9 +2,9 @@ import React from 'react';
 import { GrafanaTheme2 } from '@grafana/data';
 import { Card, Link, LinkButton, useStyles2 } from '@grafana/ui';
 import { css } from '@emotion/css';
-import { generateRoutePath } from 'routes/utils';
 
 import { type ExtendedProbe, type Label, ROUTES } from 'types';
+import { generateRoutePath } from 'routing/utils';
 import { useCanEditProbe } from 'hooks/useCanEditProbe';
 import { SuccessRateGaugeProbe } from 'components/Gauges';
