+++ src/components/SceneRedirecter.tsx
@@ -1,9 +1,9 @@
 import React from 'react';
 import { Navigate } from 'react-router-dom-v5-compat';
 import { LoadingPlaceholder } from '@grafana/ui';
-import { generateRoutePath } from 'routes/utils';
 
 import { ROUTES } from 'types';
+import { generateRoutePath } from 'routing/utils';
 import { useChecks } from 'data/useChecks';
 import { useQuery } from 'hooks/useQuery';
