+++ src/contexts/SMDatasourceContext.tsx
@@ -1,10 +1,10 @@
 import React, { createContext, PropsWithChildren, useContext } from 'react';
 import { PluginPage } from '@grafana/runtime';
 
+import { UninitialisedRouter } from 'routing/Routers';
 import { SMDataSource } from 'datasource/DataSource';
 import { useGetSMDatasource } from 'data/useSMSetup';
 import { CenteredSpinner } from 'components/CenteredSpinner';
-import { UninitialisedRouter } from 'components/Routing';
 
 type SMDatasourceContextValue = {
   smDS: SMDataSource;
