+++ src/page/EditCheck/EditCheck.tsx
@@ -3,10 +3,10 @@ import { useParams } from 'react-router-dom-v5-compat';
 import { Alert, Button, LinkButton, Modal } from '@grafana/ui';
 
 import { CheckPageParams, ROUTES } from 'types';
+import { getRoute } from 'routing/utils';
 import { useChecks } from 'data/useChecks';
 import { useNavigation } from 'hooks/useNavigation';
 import { CheckForm } from 'components/CheckForm/CheckForm';
-import { getRoute } from 'components/Routing.utils';
 
 export const EditCheck = () => {
   return <EditCheckContent />;
