+++ src/page/NewProbe/NewProbe.test.tsx
@@ -6,9 +6,8 @@ import { render } from 'test/render';
 import { fillProbeForm } from 'test/utils';
 
 import { FeatureName, ROUTES } from 'types';
-import { getRoute } from 'components/Routing.utils';
+import { generateRoutePath, getRoute } from 'routing/utils';
 
-import { generateRoutePath } from '../../routes';
 import { DataTestIds } from '../../test/dataTestIds';
 import { NewProbe } from './NewProbe';
