+++ src/page/NotFound/CheckNotFound.tsx
@@ -1,10 +1,10 @@
 import React from 'react';
 import { useParams } from 'react-router-dom-v5-compat';
 import { TextLink } from '@grafana/ui';
-import { generateRoutePath } from 'routes/utils';
 
 import { ROUTES } from 'types';
 import { createNavModel } from 'utils';
+import { generateRoutePath } from 'routing/utils';
 import { useCheck } from 'data/useChecks';
 
 import { NotFound, PluginPageNotFound } from './NotFound';
