+++ src/test/mocks/routing/TestRouteInfo.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { useLocation } from 'react-router-dom-v5-compat';
 
-import { DataTestIds } from '../../test/dataTestIds';
+import { DataTestIds } from '../../dataTestIds';
 
 /**
  * TestRouteInfo component.
